+++ resources/lang/sk/edu.php
@@ -15,9 +15,11 @@ return [
       'collection' => 'kolekcia',
       'workshop' => 'workshop',
       'virtual_exhibition' => 'virtuálna prehliadka',
+      'activities' => 'aktivity',
     ],
-    
+
     'age_group' => [
+      'all-ages' => 'bez obmedzenia',
       '3-6' => '3—6 rokov',
       '7-10' => '7—10 rokov',
       '11-15' => '11—15 rokov',